Refactor prompt and Jira scripts; add runbook and wrappers
