+++ RUNBOOK.md
@@ -1,0 +1,61 @@
+# Инструкция по Генерации Тестов
+
+Этот документ описывает два основных рабочих процесса: создание компонентных тестов и создание сценарных тестов.
+
+---
+
+## Workflow 1: Генерация Компонентных Тестов
+
+**Цель:** Создать детальные компонентные тесты на основе дизайна в Figma, требований и Swagger. Результат этого воркфлоу (`component_tests.csv`) является входными данными для генерации сценарных тестов.
+
+**Шаг 1: Получение тестов из Figma (если требуется обновление)**
+*   Убедитесь, что в `config.py` установлен `OPERATIONAL_MODE = "FILE_EXPORT"`.
+*   Запустите скрипт для анализа Figma:
+    ```bash
+    python3 send_figma_tests_all_tests.py
+    ```
+*   **Результат:** В папке `create_final_tests/artifacts/` будет создан или обновлен файл `tests_from_figma.csv`.
+
+**Шаг 2: Генерация промпта для AI**
+*   Запустите скрипт, который соберет все артефакты (требования, swagger, тесты из Figma) в один промпт.
+    ```bash
+    ./generate_component_prompt.sh
+    ```
+*   **Результат:** Будет создан файл `create_final_tests/artifacts/prompt_component.txt`.
+
+**Шаг 3: Работа с AI и получение JSON**
+1.  Скопируйте всё содержимое файла `prompt_component.txt` и передайте его AI.
+2.  Полученный от AI ответ в формате JSON сохраните в файл: `create_final_tests/artifacts/component_tests.json`.
+
+**Шаг 4: Отправка в Jira и/или конвертация в CSV**
+*   Запустите скрипт, чтобы отправить созданные компонентные тесты в Jira.
+    ```bash
+    python3 create_final_tests/jira_sender.py --input create_final_tests/artifacts/component_tests.json
+    ```
+*   *Примечание: Чтобы использовать эти тесты в следующем workflow, необходимо будет вручную или с помощью скрипта конвертировать `component_tests.json` в формат `component_tests.csv`.*
+
+---
+
+## Workflow 2: Генерация Сценарных Тестов
+
+**Цель:** Создать высокоуровневые end-to-end сценарии на основе уже готовых компонентных тестов, требований и Swagger.
+
+**Обязательное условие:** У вас должен быть актуальный файл `create_final_tests/artifacts/component_tests.csv`. Он создается на предыдущем этапе.
+
+**Шаг 1: Генерация промпта для AI**
+*   Запустите скрипт, который соберет все необходимые артефакты.
+    ```bash
+    ./generate_scenario_prompt.sh
+    ```
+*   **Результат:** Будет создан файл `create_final_tests/artifacts/prompt_scenario.txt`.
+
+**Шаг 2: Работа с AI и получение JSON**
+1.  Скопируйте всё содержимое файла `prompt_scenario.txt` и передайте его AI.
+2.  Полученный от AI ответ в формате JSON сохраните в файл: `create_final_tests/artifacts/final_tests.json`.
+
+**Шаг 3: Отправка сценарных тестов в Jira**
+*   Запустите скрипт для отправки тестов.
+    ```bash
+    python3 send_final_tests.py
+    ```
+*   **Результат:** Сценарные тесты будут созданы в Jira. Ссылка на созданные задачи появится в консоли.
