+++ create_final_tests/config_artifacts_scenario.json
@@ -1,5 +1,5 @@
 {
-  "prompt_template_path": "artifacts/prompt.md",
+  "prompt_template_path": "artifacts/prompt_scenario.md",
   "output_prompt_path": "artifacts/final_prompt.txt",
   "artifacts": {
     "req": "artifacts/req.md",
