+++ create_final_tests/prompt_generator.py
@@ -1,28 +1,25 @@
 import json
 from pathlib import Path
-import os
 import config as config_module
 
 
-def main():
+def generate_prompt_from_config(config_path: str, output_path: str) -> bool:
+    """Generate a prompt file based on a JSON config of artifacts."""
     script_dir = Path(__file__).resolve().parent
-    config_path = script_dir / 'config_artifacts.json'
-
     try:
         with open(config_path, 'r', encoding='utf-8') as f:
             config_data = json.load(f)
     except FileNotFoundError:
         print(f"❌ Ошибка: Конфигурационный файл '{config_path}' не найден.")
-        return
+        return False
     except json.JSONDecodeError:
         print(f"❌ Ошибка: Некорректный формат JSON в файле '{config_path}'.")
-        return
+        return False
     except Exception as e:
         print(f"❌ Ошибка при чтении конфигурационного файла '{config_path}': {e}")
-        return
+        return False
 
     prompt_template_path = config_data.get('prompt_template_path')
-    output_prompt_path = config_data.get('output_prompt_path')
     artifacts_paths = config_data.get('artifacts', {})
     placeholders = config_data.get('placeholders', {})
 
@@ -36,31 +33,26 @@ def main():
 
     if not prompt_template_path:
         print("❌ Ошибка: 'prompt_template_path' не указан в конфигурации.")
-        return
-    if not output_prompt_path:
-        print("❌ Ошибка: 'output_prompt_path' не указан в конфигурации.")
-        return
+        return False
 
-    if prompt_template_path:
-        path_obj = Path(prompt_template_path)
-        if not path_obj.is_absolute():
-            path_obj = script_dir / path_obj
-        prompt_template_path = path_obj
-    if output_prompt_path:
-        path_obj = Path(output_prompt_path)
-        if not path_obj.is_absolute():
-            path_obj = script_dir / path_obj
-        output_prompt_path = path_obj
+    path_obj = Path(prompt_template_path)
+    if not path_obj.is_absolute():
+        path_obj = script_dir / path_obj
+    prompt_template_path = path_obj
+
+    output_prompt_path = Path(output_path)
+    if not output_prompt_path.is_absolute():
+        output_prompt_path = script_dir / output_prompt_path
 
     try:
         with open(prompt_template_path, 'r', encoding='utf-8') as f:
             prompt_content = f.read()
     except FileNotFoundError:
         print(f"❌ Ошибка: Файл шаблона '{prompt_template_path}' не найден.")
-        return
+        return False
     except Exception as e:
         print(f"❌ Ошибка при чтении файла шаблона '{prompt_template_path}': {e}")
-        return
+        return False
 
     for artifact_key, artifact_path in artifacts_paths.items():
         placeholder = placeholders.get(artifact_key)
@@ -82,10 +74,10 @@ def main():
             print(
                 f"❌ Ошибка: Файл артефакта '{artifact_path}' (для ключа '{artifact_key}') не найден. Выполнение прервано."
             )
-            return
+            return False
         except Exception as e:
             print(f"❌ Ошибка при чтении файла артефакта '{artifact_path}': {e}")
-            return
+            return False
 
     swagger_url = getattr(config_module, 'SWAGGER_URL', '')
     prompt_content = prompt_content.replace('{{SWAGGER_URL}}', swagger_url)
@@ -94,9 +86,21 @@ def main():
         with open(output_prompt_path, 'w', encoding='utf-8') as f:
             f.write(prompt_content)
         print(f"✅ Файл '{output_prompt_path}' успешно сгенерирован.")
+        return True
     except Exception as e:
         print(f"❌ Ошибка при записи в файл '{output_prompt_path}': {e}")
+        return False
 
 
 if __name__ == '__main__':
-    main()
+    import argparse
+    parser = argparse.ArgumentParser(description="Generate a final prompt from a template and artifacts.")
+    parser.add_argument("--config", required=True, help="Path to the JSON config file for artifacts.")
+    parser.add_argument("--output", required=True, help="Path to the output prompt file.")
+    args = parser.parse_args()
+
+    if generate_prompt_from_config(args.config, args.output):
+        print(f"✅ Prompt successfully generated at: {args.output}")
+    else:
+        print(f"❌ Failed to generate prompt.")
+        exit(1)
